+++ server/weatherClients/currentWeather.go
