+++ server/weatherClients/exampleResponse.go
@@ -1,3 +1,3 @@
-package weatherAPIClient
+package weatherClients
 
 var ExampleResponse []byte = []byte("{\"data\":{\"timelines\":[{\"timestep\":\"1m\",\"startTime\":\"2021-07-29T21:51:00-04:00\",\"endTime\":\"2021-07-30T03:51:00-04:00\",\"intervals\":[{\"startTime\":\"2021-07-29T21:51:00-04:00\",\"values\":{\"temperature\":79.02,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T21:52:00-04:00\",\"values\":{\"temperature\":79.23,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T21:53:00-04:00\",\"values\":{\"temperature\":79.43,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T21:54:00-04:00\",\"values\":{\"temperature\":79.65,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T21:55:00-04:00\",\"values\":{\"temperature\":79.86,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T21:56:00-04:00\",\"values\":{\"temperature\":80.08,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T21:57:00-04:00\",\"values\":{\"temperature\":80.28,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T21:58:00-04:00\",\"values\":{\"temperature\":80.49,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T21:59:00-04:00\",\"values\":{\"temperature\":80.71,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:00:00-04:00\",\"values\":{\"temperature\":80.92,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:01:00-04:00\",\"values\":{\"temperature\":80.92,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:02:00-04:00\",\"values\":{\"temperature\":80.92,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:03:00-04:00\",\"values\":{\"temperature\":80.91,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:04:00-04:00\",\"values\":{\"temperature\":80.91,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:05:00-04:00\",\"values\":{\"temperature\":80.91,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:06:00-04:00\",\"values\":{\"temperature\":80.91,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:07:00-04:00\",\"values\":{\"temperature\":80.91,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:08:00-04:00\",\"values\":{\"temperature\":80.91,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:09:00-04:00\",\"values\":{\"temperature\":80.91,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:10:00-04:00\",\"values\":{\"temperature\":80.91,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:11:00-04:00\",\"values\":{\"temperature\":80.91,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:12:00-04:00\",\"values\":{\"temperature\":80.89,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:13:00-04:00\",\"values\":{\"temperature\":80.89,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:14:00-04:00\",\"values\":{\"temperature\":80.89,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:15:00-04:00\",\"values\":{\"temperature\":80.89,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:16:00-04:00\",\"values\":{\"temperature\":80.89,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:17:00-04:00\",\"values\":{\"temperature\":80.89,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:18:00-04:00\",\"values\":{\"temperature\":80.89,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:19:00-04:00\",\"values\":{\"temperature\":80.89,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:20:00-04:00\",\"values\":{\"temperature\":80.89,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:21:00-04:00\",\"values\":{\"temperature\":80.87,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:22:00-04:00\",\"values\":{\"temperature\":80.87,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:23:00-04:00\",\"values\":{\"temperature\":80.87,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:24:00-04:00\",\"values\":{\"temperature\":80.87,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:25:00-04:00\",\"values\":{\"temperature\":80.87,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:26:00-04:00\",\"values\":{\"temperature\":80.87,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:27:00-04:00\",\"values\":{\"temperature\":80.87,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:28:00-04:00\",\"values\":{\"temperature\":80.87,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:29:00-04:00\",\"values\":{\"temperature\":80.87,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:30:00-04:00\",\"values\":{\"temperature\":80.85,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:31:00-04:00\",\"values\":{\"temperature\":80.85,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:32:00-04:00\",\"values\":{\"temperature\":80.85,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:33:00-04:00\",\"values\":{\"temperature\":80.85,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:34:00-04:00\",\"values\":{\"temperature\":80.85,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:35:00-04:00\",\"values\":{\"temperature\":80.85,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:36:00-04:00\",\"values\":{\"temperature\":80.85,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:37:00-04:00\",\"values\":{\"temperature\":80.85,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:38:00-04:00\",\"values\":{\"temperature\":80.85,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:39:00-04:00\",\"values\":{\"temperature\":80.85,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:40:00-04:00\",\"values\":{\"temperature\":80.83,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:41:00-04:00\",\"values\":{\"temperature\":80.83,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:42:00-04:00\",\"values\":{\"temperature\":80.83,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:43:00-04:00\",\"values\":{\"temperature\":80.83,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:44:00-04:00\",\"values\":{\"temperature\":80.83,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:45:00-04:00\",\"values\":{\"temperature\":80.83,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:46:00-04:00\",\"values\":{\"temperature\":80.83,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:47:00-04:00\",\"values\":{\"temperature\":80.83,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:48:00-04:00\",\"values\":{\"temperature\":80.83,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:49:00-04:00\",\"values\":{\"temperature\":80.82,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:50:00-04:00\",\"values\":{\"temperature\":80.82,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:51:00-04:00\",\"values\":{\"temperature\":80.82,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:52:00-04:00\",\"values\":{\"temperature\":80.82,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:53:00-04:00\",\"values\":{\"temperature\":80.82,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:54:00-04:00\",\"values\":{\"temperature\":80.82,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:55:00-04:00\",\"values\":{\"temperature\":80.82,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:56:00-04:00\",\"values\":{\"temperature\":80.82,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:57:00-04:00\",\"values\":{\"temperature\":80.82,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:58:00-04:00\",\"values\":{\"temperature\":80.8,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T22:59:00-04:00\",\"values\":{\"temperature\":80.8,\"precipitationType\":1,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:00:00-04:00\",\"values\":{\"temperature\":80.8,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:01:00-04:00\",\"values\":{\"temperature\":80.8,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:02:00-04:00\",\"values\":{\"temperature\":80.78,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:03:00-04:00\",\"values\":{\"temperature\":80.78,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:04:00-04:00\",\"values\":{\"temperature\":80.76,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:05:00-04:00\",\"values\":{\"temperature\":80.76,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:06:00-04:00\",\"values\":{\"temperature\":80.74,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:07:00-04:00\",\"values\":{\"temperature\":80.73,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:08:00-04:00\",\"values\":{\"temperature\":80.73,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:09:00-04:00\",\"values\":{\"temperature\":80.71,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:10:00-04:00\",\"values\":{\"temperature\":80.71,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:11:00-04:00\",\"values\":{\"temperature\":80.69,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:12:00-04:00\",\"values\":{\"temperature\":80.69,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:13:00-04:00\",\"values\":{\"temperature\":80.67,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:14:00-04:00\",\"values\":{\"temperature\":80.67,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:15:00-04:00\",\"values\":{\"temperature\":80.65,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:16:00-04:00\",\"values\":{\"temperature\":80.65,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:17:00-04:00\",\"values\":{\"temperature\":80.64,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:18:00-04:00\",\"values\":{\"temperature\":80.62,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:19:00-04:00\",\"values\":{\"temperature\":80.62,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:20:00-04:00\",\"values\":{\"temperature\":80.6,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:21:00-04:00\",\"values\":{\"temperature\":80.6,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:22:00-04:00\",\"values\":{\"temperature\":80.58,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:23:00-04:00\",\"values\":{\"temperature\":80.58,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:24:00-04:00\",\"values\":{\"temperature\":80.56,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:25:00-04:00\",\"values\":{\"temperature\":80.56,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:26:00-04:00\",\"values\":{\"temperature\":80.55,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-29T23:27:00-04:00\",\"values\":{\"temperature\":80.55,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:28:00-04:00\",\"values\":{\"temperature\":80.53,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:29:00-04:00\",\"values\":{\"temperature\":80.53,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:30:00-04:00\",\"values\":{\"temperature\":80.51,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:31:00-04:00\",\"values\":{\"temperature\":80.49,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:32:00-04:00\",\"values\":{\"temperature\":80.49,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:33:00-04:00\",\"values\":{\"temperature\":80.47,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:34:00-04:00\",\"values\":{\"temperature\":80.47,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:35:00-04:00\",\"values\":{\"temperature\":80.46,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:36:00-04:00\",\"values\":{\"temperature\":80.46,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:37:00-04:00\",\"values\":{\"temperature\":80.44,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:38:00-04:00\",\"values\":{\"temperature\":80.44,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:39:00-04:00\",\"values\":{\"temperature\":80.42,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:40:00-04:00\",\"values\":{\"temperature\":80.42,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-29T23:41:00-04:00\",\"values\":{\"temperature\":80.4,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:42:00-04:00\",\"values\":{\"temperature\":80.38,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:43:00-04:00\",\"values\":{\"temperature\":80.38,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:44:00-04:00\",\"values\":{\"temperature\":80.37,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:45:00-04:00\",\"values\":{\"temperature\":80.37,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:46:00-04:00\",\"values\":{\"temperature\":80.35,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:47:00-04:00\",\"values\":{\"temperature\":80.35,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:48:00-04:00\",\"values\":{\"temperature\":80.33,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:49:00-04:00\",\"values\":{\"temperature\":80.33,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:50:00-04:00\",\"values\":{\"temperature\":80.31,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:51:00-04:00\",\"values\":{\"temperature\":80.31,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:52:00-04:00\",\"values\":{\"temperature\":80.29,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:53:00-04:00\",\"values\":{\"temperature\":80.29,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:54:00-04:00\",\"values\":{\"temperature\":80.28,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:55:00-04:00\",\"values\":{\"temperature\":80.26,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:56:00-04:00\",\"values\":{\"temperature\":80.26,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:57:00-04:00\",\"values\":{\"temperature\":80.24,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:58:00-04:00\",\"values\":{\"temperature\":80.24,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-29T23:59:00-04:00\",\"values\":{\"temperature\":80.22,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:00:00-04:00\",\"values\":{\"temperature\":80.22,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:01:00-04:00\",\"values\":{\"temperature\":80.2,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:02:00-04:00\",\"values\":{\"temperature\":80.2,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:03:00-04:00\",\"values\":{\"temperature\":80.2,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:04:00-04:00\",\"values\":{\"temperature\":80.19,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:05:00-04:00\",\"values\":{\"temperature\":80.19,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:06:00-04:00\",\"values\":{\"temperature\":80.19,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:07:00-04:00\",\"values\":{\"temperature\":80.17,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:08:00-04:00\",\"values\":{\"temperature\":80.17,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:09:00-04:00\",\"values\":{\"temperature\":80.17,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:10:00-04:00\",\"values\":{\"temperature\":80.15,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:11:00-04:00\",\"values\":{\"temperature\":80.15,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:12:00-04:00\",\"values\":{\"temperature\":80.15,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:13:00-04:00\",\"values\":{\"temperature\":80.13,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:14:00-04:00\",\"values\":{\"temperature\":80.13,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:15:00-04:00\",\"values\":{\"temperature\":80.11,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:16:00-04:00\",\"values\":{\"temperature\":80.11,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:17:00-04:00\",\"values\":{\"temperature\":80.11,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:18:00-04:00\",\"values\":{\"temperature\":80.1,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:19:00-04:00\",\"values\":{\"temperature\":80.1,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:20:00-04:00\",\"values\":{\"temperature\":80.1,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:21:00-04:00\",\"values\":{\"temperature\":80.08,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:22:00-04:00\",\"values\":{\"temperature\":80.08,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:23:00-04:00\",\"values\":{\"temperature\":80.08,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:24:00-04:00\",\"values\":{\"temperature\":80.06,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:25:00-04:00\",\"values\":{\"temperature\":80.06,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:26:00-04:00\",\"values\":{\"temperature\":80.06,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:27:00-04:00\",\"values\":{\"temperature\":80.04,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:28:00-04:00\",\"values\":{\"temperature\":80.04,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:29:00-04:00\",\"values\":{\"temperature\":80.04,\"precipitationType\":0,\"weatherCode\":1101}},{\"startTime\":\"2021-07-30T00:30:00-04:00\",\"values\":{\"temperature\":80.02,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:31:00-04:00\",\"values\":{\"temperature\":80.02,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:32:00-04:00\",\"values\":{\"temperature\":80.01,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:33:00-04:00\",\"values\":{\"temperature\":80.01,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:34:00-04:00\",\"values\":{\"temperature\":80.01,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:35:00-04:00\",\"values\":{\"temperature\":79.99,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:36:00-04:00\",\"values\":{\"temperature\":79.99,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:37:00-04:00\",\"values\":{\"temperature\":79.99,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:38:00-04:00\",\"values\":{\"temperature\":79.97,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:39:00-04:00\",\"values\":{\"temperature\":79.97,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:40:00-04:00\",\"values\":{\"temperature\":79.97,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:41:00-04:00\",\"values\":{\"temperature\":79.95,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:42:00-04:00\",\"values\":{\"temperature\":79.95,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:43:00-04:00\",\"values\":{\"temperature\":79.95,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:44:00-04:00\",\"values\":{\"temperature\":79.93,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:45:00-04:00\",\"values\":{\"temperature\":79.93,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:46:00-04:00\",\"values\":{\"temperature\":79.93,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:47:00-04:00\",\"values\":{\"temperature\":79.92,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:48:00-04:00\",\"values\":{\"temperature\":79.92,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:49:00-04:00\",\"values\":{\"temperature\":79.92,\"precipitationType\":0,\"weatherCode\":1102}},{\"startTime\":\"2021-07-30T00:50:00-04:00\",\"values\":{\"temperature\":79.9,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T00:51:00-04:00\",\"values\":{\"temperature\":79.9,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T00:52:00-04:00\",\"values\":{\"temperature\":79.88,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T00:53:00-04:00\",\"values\":{\"temperature\":79.88,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T00:54:00-04:00\",\"values\":{\"temperature\":79.88,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T00:55:00-04:00\",\"values\":{\"temperature\":79.86,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T00:56:00-04:00\",\"values\":{\"temperature\":79.86,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T00:57:00-04:00\",\"values\":{\"temperature\":79.86,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T00:58:00-04:00\",\"values\":{\"temperature\":79.84,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T00:59:00-04:00\",\"values\":{\"temperature\":79.84,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:00:00-04:00\",\"values\":{\"temperature\":79.84,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:01:00-04:00\",\"values\":{\"temperature\":79.83,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:02:00-04:00\",\"values\":{\"temperature\":79.83,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:03:00-04:00\",\"values\":{\"temperature\":79.81,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:04:00-04:00\",\"values\":{\"temperature\":79.79,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:05:00-04:00\",\"values\":{\"temperature\":79.79,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:06:00-04:00\",\"values\":{\"temperature\":79.77,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:07:00-04:00\",\"values\":{\"temperature\":79.77,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:08:00-04:00\",\"values\":{\"temperature\":79.75,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:09:00-04:00\",\"values\":{\"temperature\":79.75,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:10:00-04:00\",\"values\":{\"temperature\":79.74,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:11:00-04:00\",\"values\":{\"temperature\":79.72,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:12:00-04:00\",\"values\":{\"temperature\":79.72,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:13:00-04:00\",\"values\":{\"temperature\":79.7,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:14:00-04:00\",\"values\":{\"temperature\":79.7,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:15:00-04:00\",\"values\":{\"temperature\":79.68,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:16:00-04:00\",\"values\":{\"temperature\":79.68,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:17:00-04:00\",\"values\":{\"temperature\":79.66,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:18:00-04:00\",\"values\":{\"temperature\":79.65,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:19:00-04:00\",\"values\":{\"temperature\":79.65,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:20:00-04:00\",\"values\":{\"temperature\":79.63,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:21:00-04:00\",\"values\":{\"temperature\":79.63,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:22:00-04:00\",\"values\":{\"temperature\":79.61,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:23:00-04:00\",\"values\":{\"temperature\":79.61,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:24:00-04:00\",\"values\":{\"temperature\":79.59,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:25:00-04:00\",\"values\":{\"temperature\":79.57,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:26:00-04:00\",\"values\":{\"temperature\":79.57,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:27:00-04:00\",\"values\":{\"temperature\":79.56,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:28:00-04:00\",\"values\":{\"temperature\":79.56,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:29:00-04:00\",\"values\":{\"temperature\":79.54,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:30:00-04:00\",\"values\":{\"temperature\":79.54,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:31:00-04:00\",\"values\":{\"temperature\":79.52,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:32:00-04:00\",\"values\":{\"temperature\":79.5,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:33:00-04:00\",\"values\":{\"temperature\":79.5,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:34:00-04:00\",\"values\":{\"temperature\":79.48,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:35:00-04:00\",\"values\":{\"temperature\":79.48,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:36:00-04:00\",\"values\":{\"temperature\":79.47,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:37:00-04:00\",\"values\":{\"temperature\":79.47,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:38:00-04:00\",\"values\":{\"temperature\":79.45,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:39:00-04:00\",\"values\":{\"temperature\":79.43,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:40:00-04:00\",\"values\":{\"temperature\":79.43,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:41:00-04:00\",\"values\":{\"temperature\":79.41,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:42:00-04:00\",\"values\":{\"temperature\":79.41,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:43:00-04:00\",\"values\":{\"temperature\":79.39,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:44:00-04:00\",\"values\":{\"temperature\":79.39,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:45:00-04:00\",\"values\":{\"temperature\":79.38,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:46:00-04:00\",\"values\":{\"temperature\":79.36,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:47:00-04:00\",\"values\":{\"temperature\":79.36,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:48:00-04:00\",\"values\":{\"temperature\":79.34,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:49:00-04:00\",\"values\":{\"temperature\":79.34,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:50:00-04:00\",\"values\":{\"temperature\":79.32,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:51:00-04:00\",\"values\":{\"temperature\":79.32,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:52:00-04:00\",\"values\":{\"temperature\":79.3,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:53:00-04:00\",\"values\":{\"temperature\":79.29,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:54:00-04:00\",\"values\":{\"temperature\":79.29,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:55:00-04:00\",\"values\":{\"temperature\":79.27,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:56:00-04:00\",\"values\":{\"temperature\":79.27,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:57:00-04:00\",\"values\":{\"temperature\":79.25,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:58:00-04:00\",\"values\":{\"temperature\":79.25,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T01:59:00-04:00\",\"values\":{\"temperature\":79.23,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:00:00-04:00\",\"values\":{\"temperature\":79.21,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:01:00-04:00\",\"values\":{\"temperature\":79.21,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:02:00-04:00\",\"values\":{\"temperature\":79.2,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:03:00-04:00\",\"values\":{\"temperature\":79.18,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:04:00-04:00\",\"values\":{\"temperature\":79.18,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:05:00-04:00\",\"values\":{\"temperature\":79.16,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:06:00-04:00\",\"values\":{\"temperature\":79.14,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:07:00-04:00\",\"values\":{\"temperature\":79.14,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:08:00-04:00\",\"values\":{\"temperature\":79.12,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:09:00-04:00\",\"values\":{\"temperature\":79.11,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:10:00-04:00\",\"values\":{\"temperature\":79.11,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:11:00-04:00\",\"values\":{\"temperature\":79.09,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:12:00-04:00\",\"values\":{\"temperature\":79.07,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:13:00-04:00\",\"values\":{\"temperature\":79.07,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:14:00-04:00\",\"values\":{\"temperature\":79.05,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:15:00-04:00\",\"values\":{\"temperature\":79.03,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:16:00-04:00\",\"values\":{\"temperature\":79.03,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:17:00-04:00\",\"values\":{\"temperature\":79.02,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:18:00-04:00\",\"values\":{\"temperature\":79,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:19:00-04:00\",\"values\":{\"temperature\":79,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:20:00-04:00\",\"values\":{\"temperature\":78.98,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:21:00-04:00\",\"values\":{\"temperature\":78.96,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:22:00-04:00\",\"values\":{\"temperature\":78.96,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:23:00-04:00\",\"values\":{\"temperature\":78.94,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:24:00-04:00\",\"values\":{\"temperature\":78.94,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:25:00-04:00\",\"values\":{\"temperature\":78.93,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:26:00-04:00\",\"values\":{\"temperature\":78.91,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:27:00-04:00\",\"values\":{\"temperature\":78.91,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:28:00-04:00\",\"values\":{\"temperature\":78.89,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:29:00-04:00\",\"values\":{\"temperature\":78.87,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:30:00-04:00\",\"values\":{\"temperature\":78.87,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:31:00-04:00\",\"values\":{\"temperature\":78.85,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:32:00-04:00\",\"values\":{\"temperature\":78.84,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:33:00-04:00\",\"values\":{\"temperature\":78.84,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:34:00-04:00\",\"values\":{\"temperature\":78.82,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:35:00-04:00\",\"values\":{\"temperature\":78.8,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:36:00-04:00\",\"values\":{\"temperature\":78.8,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:37:00-04:00\",\"values\":{\"temperature\":78.78,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:38:00-04:00\",\"values\":{\"temperature\":78.76,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:39:00-04:00\",\"values\":{\"temperature\":78.76,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:40:00-04:00\",\"values\":{\"temperature\":78.75,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:41:00-04:00\",\"values\":{\"temperature\":78.73,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:42:00-04:00\",\"values\":{\"temperature\":78.73,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:43:00-04:00\",\"values\":{\"temperature\":78.71,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:44:00-04:00\",\"values\":{\"temperature\":78.69,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:45:00-04:00\",\"values\":{\"temperature\":78.69,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:46:00-04:00\",\"values\":{\"temperature\":78.67,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:47:00-04:00\",\"values\":{\"temperature\":78.66,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:48:00-04:00\",\"values\":{\"temperature\":78.66,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:49:00-04:00\",\"values\":{\"temperature\":78.64,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:50:00-04:00\",\"values\":{\"temperature\":78.62,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:51:00-04:00\",\"values\":{\"temperature\":78.62,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:52:00-04:00\",\"values\":{\"temperature\":78.6,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:53:00-04:00\",\"values\":{\"temperature\":78.58,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:54:00-04:00\",\"values\":{\"temperature\":78.58,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:55:00-04:00\",\"values\":{\"temperature\":78.57,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:56:00-04:00\",\"values\":{\"temperature\":78.55,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:57:00-04:00\",\"values\":{\"temperature\":78.55,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:58:00-04:00\",\"values\":{\"temperature\":78.53,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T02:59:00-04:00\",\"values\":{\"temperature\":78.51,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:00:00-04:00\",\"values\":{\"temperature\":78.51,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:01:00-04:00\",\"values\":{\"temperature\":78.49,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:02:00-04:00\",\"values\":{\"temperature\":78.49,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:03:00-04:00\",\"values\":{\"temperature\":78.49,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:04:00-04:00\",\"values\":{\"temperature\":78.48,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:05:00-04:00\",\"values\":{\"temperature\":78.48,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:06:00-04:00\",\"values\":{\"temperature\":78.48,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:07:00-04:00\",\"values\":{\"temperature\":78.46,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:08:00-04:00\",\"values\":{\"temperature\":78.46,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:09:00-04:00\",\"values\":{\"temperature\":78.46,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:10:00-04:00\",\"values\":{\"temperature\":78.46,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:11:00-04:00\",\"values\":{\"temperature\":78.44,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:12:00-04:00\",\"values\":{\"temperature\":78.44,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:13:00-04:00\",\"values\":{\"temperature\":78.44,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:14:00-04:00\",\"values\":{\"temperature\":78.42,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:15:00-04:00\",\"values\":{\"temperature\":78.42,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:16:00-04:00\",\"values\":{\"temperature\":78.42,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:17:00-04:00\",\"values\":{\"temperature\":78.4,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:18:00-04:00\",\"values\":{\"temperature\":78.4,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:19:00-04:00\",\"values\":{\"temperature\":78.4,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:20:00-04:00\",\"values\":{\"temperature\":78.39,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:21:00-04:00\",\"values\":{\"temperature\":78.39,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:22:00-04:00\",\"values\":{\"temperature\":78.39,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:23:00-04:00\",\"values\":{\"temperature\":78.37,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:24:00-04:00\",\"values\":{\"temperature\":78.37,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:25:00-04:00\",\"values\":{\"temperature\":78.37,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:26:00-04:00\",\"values\":{\"temperature\":78.35,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:27:00-04:00\",\"values\":{\"temperature\":78.35,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:28:00-04:00\",\"values\":{\"temperature\":78.35,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:29:00-04:00\",\"values\":{\"temperature\":78.33,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:30:00-04:00\",\"values\":{\"temperature\":78.33,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:31:00-04:00\",\"values\":{\"temperature\":78.33,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:32:00-04:00\",\"values\":{\"temperature\":78.31,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:33:00-04:00\",\"values\":{\"temperature\":78.31,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:34:00-04:00\",\"values\":{\"temperature\":78.31,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:35:00-04:00\",\"values\":{\"temperature\":78.3,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:36:00-04:00\",\"values\":{\"temperature\":78.3,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:37:00-04:00\",\"values\":{\"temperature\":78.3,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:38:00-04:00\",\"values\":{\"temperature\":78.28,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:39:00-04:00\",\"values\":{\"temperature\":78.28,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:40:00-04:00\",\"values\":{\"temperature\":78.28,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:41:00-04:00\",\"values\":{\"temperature\":78.26,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:42:00-04:00\",\"values\":{\"temperature\":78.26,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:43:00-04:00\",\"values\":{\"temperature\":78.26,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:44:00-04:00\",\"values\":{\"temperature\":78.24,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:45:00-04:00\",\"values\":{\"temperature\":78.24,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:46:00-04:00\",\"values\":{\"temperature\":78.24,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:47:00-04:00\",\"values\":{\"temperature\":78.24,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:48:00-04:00\",\"values\":{\"temperature\":78.22,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:49:00-04:00\",\"values\":{\"temperature\":78.22,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:50:00-04:00\",\"values\":{\"temperature\":78.22,\"precipitationType\":0,\"weatherCode\":1001}},{\"startTime\":\"2021-07-30T03:51:00-04:00\",\"values\":{\"temperature\":78.21,\"precipitationType\":0,\"weatherCode\":1001}}]}]}}")
